refactor: TermsValidator 리팩토링 및 구현체 삭제 - interface에서 class로 수정 - 예외 메시지 상수로 변경 - 메서드 분리
